+++ modules/src/ics03_connection/msgs.rs
@@ -11,6 +11,7 @@
 //!
 //! Another difference to ICS3 specs is that each message comprises an additional field called
 //! `signer` which is specific to Cosmos-SDK.
+
 use crate::ics03_connection::msgs::conn_open_ack::MsgConnectionOpenAck;
 use crate::ics03_connection::msgs::conn_open_confirm::MsgConnectionOpenConfirm;
 use crate::ics03_connection::msgs::conn_open_init::MsgConnectionOpenInit;
@@ -21,8 +22,14 @@ pub mod conn_open_confirm;
 pub mod conn_open_init;
 pub mod conn_open_try;
 
-/// Message type for the `MsgConnectionOpenConfirm` message.
-pub const TYPE_MSG_CONNECTION_OPEN_CONFIRM: &str = "connection_open_confirm";
+/// Enumeration of all possible message types that the ICS3 protocol processes.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub enum ConnectionMsgType {
+    OpenInit,
+    OpenTry,
+    OpenAck,
+    OpenConfirm,
+}
 
 /// Enumeration of all possible messages that the ICS3 protocol processes.
 #[derive(Clone, Debug, PartialEq, Eq)]