+++ modules/src/mock/context.rs
@@ -279,17 +279,6 @@ impl ConnectionReader for MockContext {
         let block_ref = self.host_block(height);
         block_ref.cloned().map(Into::into)
     }
-
-    fn get_compatible_versions(&self) -> Vec<String> {
-        vec!["test".to_string()]
-    }
-
-    fn pick_version(&self, counterparty_candidate_versions: Vec<String>) -> String {
-        counterparty_candidate_versions
-            .get(0)
-            .unwrap_or(&String::from("none"))
-            .to_string()
-    }
 }
 
 impl ConnectionKeeper for MockContext {