+++ relayer-cli/src/commands/query/connection.rs
@@ -1,9 +1,7 @@
-use std::convert::TryInto;
-
 use abscissa_core::{Command, Options, Runnable};
 use ibc::ics03_connection::connection::ConnectionEnd;
 use ibc::ics24_host::error::ValidationError;
-use ibc::ics24_host::identifier::ConnectionId;
+use ibc::ics24_host::identifier::{ChainId as ICSChainId, ConnectionId};
 use relayer::chain::{Chain, CosmosSDKChain};
 use relayer::config::{ChainConfig, Config};
 use tendermint::chain::Id as ChainId;
@@ -64,6 +62,7 @@ impl QueryConnectionEndCmd {
     }
 }
 
+// cargo run --bin relayer -- -c relayer/tests/config/fixtures/simple_config.toml query connection end ibc-test connectionidone --height 3
 impl Runnable for QueryConnectionEndCmd {
     fn run(&self) {
         let config = app_config();
@@ -78,14 +77,14 @@ impl Runnable for QueryConnectionEndCmd {
         status_info!("Options", "{:?}", opts);
 
         let chain = CosmosSDKChain::from_config(chain_config).unwrap();
-        // run without proof:
-        // cargo run --bin relayer -- -c relayer/tests/config/fixtures/simple_config.toml query connection end ibc-test connectionidone --height 3 -p false
+        let height = ibc::Height::new(
+            ICSChainId::chain_version(chain.id().to_string()),
+            opts.height,
+        );
+
+        // TODO - any value in querying with proof from the CLI?
         let res: Result<ConnectionEnd, Error> = chain
-            .query_connection(
-                &opts.connection_id,
-                opts.height.try_into().unwrap(),
-                opts.proof,
-            )
+            .query_connection(&opts.connection_id, height)
             .map_err(|e| Kind::Query.context(e).into());
 
         match res {