+++ CHANGELOG.md
@@ -15,7 +15,9 @@ Special thanks to external contributors for this release: @CharlyCst ([#347]).
 - [relayer-cli]
     - Merge light clients config in relayer config and add commands to add/remove light clients ([#348])
     - CLI for client update message ([#277])
-- [proto-compiler] Refactor and allow specifying a commit at which the Cosmos SDK should be checked out ([#366])
+- [proto-compiler]
+    - Refactor and allow specifying a commit at which the Cosmos SDK should be checked out ([#366])
+    - Add a `--tag` option to the `clone-sdk` command to check out a tag instead of a commit ([#369])
 - [ibc-proto] Refactor and allow specifying a commit at which the Cosmos SDK should be checked out ([#366])
 
 [#274]: https://github.com/informalsystems/ibc-rs/issues/274
@@ -25,6 +27,7 @@ Special thanks to external contributors for this release: @CharlyCst ([#347]).
 [#348]: https://github.com/informalsystems/ibc-rs/pulls/348
 [#366]: https://github.com/informalsystems/ibc-rs/issues/366
 [#389]: https://github.com/informalsystems/ibc-rs/issues/389
+[#369]: https://github.com/informalsystems/ibc-rs/pulls/369
 
 ### IMPROVEMENTS